Delete -mnop-mcount from the linker
It seems that with `dune.3.7.0`, the build system try to compile
something (like a *.h) and -mnop-mcount comes from the `ld.in` but it
can be used with `-fPIC`. It's probably more a dune issue than an
esperanto issue but the purpose of mnop-mcount is not really important
when we don't use `-pg` option.